refactor: Implement find_var_type in terms of the semantic graph (#4948)
* refactor: Implement find_var_type in terms of the semantic graph

Changes `find_var_type` to just look at the semantic graph's type instead of an injected dummy variables. As part of this I had to make some changes to ensure that an undefined variable will always refer to the same thing, which may be a bit strange but I think it works fine (stop me otherwise!).

While this doesn't actually change things for `find_var_type` as it could easily just define multiple variables with one call, I'd argue this exposes less of the compiler internals to users. If we can move `find_var_type` to accept a semantic graph directly we could reduce this further.

cc #4751

* chore: make generate

* refactor: Make find_var_type accept a semantic graph

* feat: Allow multiple variables to be queried with FindVarTypes

* feat: Make FindVarTypes signature match what is used downstream

Downstream users don't care about the AST, so we can keep that internal and let them work with a string source instead.

* chore: cargo clippy