refactor: expand tests, update supported literal list
The list of literals we treat as vec repeat now:
- int
- float
- string (no interpolation)
- time

Others like duration (or regex!) are not supported. A notable omission
is uint, which would work fine but we don't have a way to write them as
literals -- we need to first vectorize the `unit()` builtin.

Had a lot of trouble getting the Go unit tests written to check the
behavior of vec repeat values.

This diff moves all the behavior checking (correctness) to flux
acceptance tests. The Go tests now simply verify that certain things
_can or cannot_ be vectorized.

N.b. equality operators are not yet implemented (#4608) but are showing
up as vectorizable. The rust code will need to be updated to prevent
this from happening until #4608 is completed.