fix(stdlib): aggregate window mean and sum work properly with null values
The native aggregate window implementations of mean and sum now properly
work with null values and the mean implementation now uses the mean of
means method of calculation rather than just `sum / count`.

The null values ensures that it properly acts when there are null values
or when the interval is only null values.

The change to use a mean of means method of calculation helps to prevent
infinite values from showing up in the result when the sum overflows
because we keep a mean of each interval with the count so that we can
merge two means.