fix(stdlib): to function should skip empty tag values
The to function is intended to skip tags that are empty since that is
how tags with an empty value are represented in influxdb.

The old implementation would also skip keys that are empty, but I think
this behavior isn't correct so this will retain the behavior of erroring
on the tag key being invalid.