fix(plan): update pattern matching to specify successor counts
The core of the problem is that the pattern matching logic
expects there to be only a single edge between a node and
its matched predecessor. This works and is required for rules
that intend to merge the matched nodes into a single node.
For example From -> Filter, if the from has other successors
you can't merge the filter into the from. In other cases this
restriction is not needed and prevents the rule for matching
when it is safe to do so. For example a join node whose predecessor
has a different successor, since the join rule doesn't rewrite its
predecessors it doesn't matter if they have other successors.

To solve this we need to be explicit about whether a pattern match
can have multiple successors or not. I propose we replace the existing
plan.Pat function with two functions plan.Multi and plan.Single.
plan.Multi will match a node and will allow that node to have multiple
successors, while the plan.Single will match a node and that node must
have exactly one successor. This changes two things about current patterns,
first it makes the number of successors expected explicit in the pattern,
second it moves the specification of which nodes can have multiple successors
and which cannot on the pattern that matches the node instead of the
successor of the node. Additionally plan.Any will allow multiple successors.

Fixes #4699