feat: add support for constants/literals in vectorized map (#4960)
* feat: add vectorRepeatValue; wrap literals with call expr, rewrite during eval

The caper is essentially to inject a fake function during the semantic
graph rewrite for vectorization, look for the function name while
evaluating call expressions, then branch to replace the node with a new
const vector evaluator (which holds a `vectorRepeatValue`).

The `vectorRepeatValue` is currently not very useful on its own. Access
to  the `Arr()` method will cause a panic, so callsites are updated to
type switch and avoid this. These callsites are where we need to branch
the behavior to account for the singleton value vs the array.

* refactor: use specialized fns, fill array from repeat in map2

Abandons the -ArrLike interfaces to use specialized vector fns for each of
left or right consts.

Also fixes the case where no binary operation produces an array-backed
vector from a repeat. When a repeat "survives" to the point of `map2`
returning, use the value to build a full array of the same value.

* refactor: constant folding now happens in binaryVectorEvaluator

By feeding the OperatorKind down through from the compiler into
binaryVectorEvaluator, we can check to see if the vectors we're
operating on are both constants (ie, `VectorRepeatValue`s).

In this case, we can short circuit the vectorized operator function and
use the regular row-based one to produce a new `VectorRepeatValue`.

* refactor: expand tests, update supported literal list

The list of literals we treat as vec repeat now:
- int
- float
- string (no interpolation)
- time

Others like duration (or regex!) are not supported. A notable omission
is uint, which would work fine but we don't have a way to write them as
literals -- we need to first vectorize the `unit()` builtin.

Had a lot of trouble getting the Go unit tests written to check the
behavior of vec repeat values.

This diff moves all the behavior checking (correctness) to flux
acceptance tests. The Go tests now simply verify that certain things
_can or cannot_ be vectorized.

N.b. equality operators are not yet implemented (#4608) but are showing
up as vectorizable. The rust code will need to be updated to prevent
this from happening until #4608 is completed.

* fix: only certain binary ops can be vectorized

Adds a check to binary expressions to prevent them from vectorizing if
the operator isn't one of the ones we've added support for.

* chore: make fmt

* refactor: move const folding into function lookup

This removes the need to attach the OperatorKind as a new field on the
evaluator since we already know what the op is while looking for the
vectorized version.

* feat: add new feature for VectorizedConst to libflux

* chore: fixup docstring

* feat: add feature flag for VectorizedConst on Go side

* refactor: move float cast outside loop for consts

* refactor: reflow map2 to reduce indent levels

* chore: make generate