feat(stdlib): add a proper diff implementation to the experimental package
This adds a proper diff transformation that computes a longest common
subsequence table and then uses that to output a diff. This is in
contrast to the existing diff in `testing.diff` that both integrates
with our testing subsystem (by using yields) and also does a naive
algorithm where it just checks if `A[X] = B[X]` instead of trying to
determine sequences that are in common.

The limitations of this implementation are as follows:

* Does not presently restrict memory usage for the LCS table. It should
  use the memory allocator to allocate the table but it constructs a
  matrix from Go slices.
* Uses a recursive algorithm to trace the LCS table instead of a
  sequential algorithm. This could cause large tables to recurse too
  much.
* Does not print only the context of a change. If it finds a difference
  in the table, it prints the entire table.