refactor: add vec repeat support for logical ops
During #4622 lots of attention was given to Vector usage in binary ops
since the literals that could be vectorized tended to show up there.

Since boolean literals were not on the list, I somehow neglected to look
at the logical operators (which depend on booleans to be used).

This diff adds the necessary branching to cover the vec repeat case,
looking forward to the time where boolean literals _can be vectorized_.

It also adds skipped/commented out tests with reference to #4997 which
should aim to enable these code paths by finally allowing bool literals
to become vectorized.