feat(stdlib): switch testing.diff to use experimental.diff as its base
This causes the `_diff` function to use the diff implementation in
`experimental` as the base diff implementation when the feature flag is
set.

This will cause diff to use a longest common subsequence algorithm to
compute the diff.

This is behind a feature flag in case it has some unknown complication.