fix(plan): modify plan.Pat to work with no predecessors for transformations
This modifies the pattern produced by `plan.Pat` so that it can work
when the topmost predecessor has predecessors but without requiring
those predecessors to be listed.

This allows plans like `sort |> limit` to work without using the `any`
node. The `any` node would cause the pattern to require the `any` node
to only have one successor, but it was perfectly allowable for that node
to have multiple successors. This makes writing the planner rule more
intuitive since we don't have to use `any` as the last predecessor and
it also prevents the problem of disallowing a pattern when it could have
been used.