feat: compute minimum required dispatcher concurrency from the plan graph
Compute the minimum dispatcher concurrency (number goroutines servicing
execution graph) in a walk of the plan graph. This is:

  sum( number of predecessors of N )
  for every node N that becomes a root of the execution graph

Any amount less than this can result in dispatcher deadlock. Details can be
found in influxdata/idpe#15220.

At this time we also add goroutines to get the desired speedup in the presence
of parallelization. We aim to add a fixed amount of goroutines based on the
parallelization factor, but take into account that some goroutines might have
been added because parallel merge nodes are also roots.

Also, we allow the new feature flag queryConcurrencyIncrease to add additional
goroutines.

The existing feature flag queryConcurrencyQuota is removed, as it is unsafe to
reduce the number of goroutines below the computed minimum.

Since the parallelization planner rules are no longer directly specifying the
desired concurrency, we can now remove the concurrency and max allocation
fields from the dependencies. This was in the context as a method of passing
concurrency requirements from a planner rule to the execution engine. The max
allocation was never required to be in the context, it just followed along with
the query concurrency so all resource limits could be defined in the same
place.