test: show that conditional expressions are vectorizable
Currently the Go unit tests that actually process inputs are failing due
to a type conflict on the `test` portion of the expression,
ex: `bool != v[bool]`.

Not sure why I didn't bump into this earlier -- manual testing via
scripts all seemed to work fine as is. Watching in the debugger showed
execution hitting all the expected (vectorized) code paths.

The "Logical Ops" PR added a `typ` field to the expression node to allow
us to alter the monotype during vectorization, so I'll do the same and
see if that fixes things.