fix: test the vectorized fn instead of the original
Fixes an issue where the tests for vectorized conditionals broke with a
type conflict on the `test` part of the expression (`bool` vs
`v[bool]`).

For whatever reason, it seems like we may have been running the
row-based version of the function prior to this diff.

Now, when a test case sets `vectorizable` to `true`, we replace the
function expression with the vectorized version before passing it to
`Compile`.