fix: add missing vec repeat handling for vectorized conditionals
Already covered were cases where `test` was const, as well as when
EITHER `consequent` OR `alternate` were const, but not BOTH.

For cases where all 3 are const, the receiver of of the return value of
the conditional expression will get the vec repeat from whichever branch
was selected (this case was already handled).

The panic from the type conflict between Time/Int is still to resolve.