refactor(promql): remove the promql package as it is not used
This package was intended to create a transpiler for the promql language
to flux. This reached a prototype stage but was never deployed as part
of a live service and hasn't been maintained since 2019.

Removing this to ensure it is not confused to be a formal part of the
flux language or Go API. This also removes a few unnecessary
dependencies which could affect downstream packages.