refactor(colm): remove the colm package
This was an experimental package about using colm with flux to transform
influxql to flux and also as an experiment for a version of flux where
the data defaulted to being "pivoted".

We learned a decent amount from this, but ultimately didn't pursue this
specific implementation. This package is unused and may be confusing for
a person who is trying to poke around the repository. Removing this so
it isn't confused as part of the Go API.