refactor(execute): transition narrow state transformation to use generics
This modifies the narrow state transformation to use generics instead of
`interface{}` for its implementation. This should allow us to write code
more easily as it will remove the boilerplate to do a type cast at the
beginning of process.

None of the transformations that use this have been modified. Instead,
they use the generic method but use `interface{}` as their type
parameter.