refactor: move lesser used packages to internal and remove unused packages (#5028)
* refactor(promql): remove the promql package as it is not used

This package was intended to create a transpiler for the promql language
to flux. This reached a prototype stage but was never deployed as part
of a live service and hasn't been maintained since 2019.

Removing this to ensure it is not confused to be a formal part of the
flux language or Go API. This also removes a few unnecessary
dependencies which could affect downstream packages.

* refactor(colm): remove the colm package

This was an experimental package about using colm with flux to transform
influxql to flux and also as an experiment for a version of flux where
the data defaulted to being "pivoted".

We learned a decent amount from this, but ultimately didn't pursue this
specific implementation. This package is unused and may be confusing for
a person who is trying to poke around the repository. Removing this so
it isn't confused as part of the Go API.

* refactor(iocounter): move iocounter package to internal

This is a utility package meant to implement a functionality for our own
code. It is not a public API. This moves it to internal to correctly
communicate that this isn't a formal part of the API contract.

* refactor(line): move the line package to internal

The line package is used to read line protocol from a socket. This isn't
currently used anywhere else. It might be useful outside of the flux
repository, but because it's a package that has low usage, it is getting
moved to internal to avoid creating an external contract we can't
necessarily guarantee outside of the socket package.

* refactor(influxql): move the influxql package to internal

This is another package that could have outside application but is
mostly used in low usage packages. In order to restrict its usage and
restrict our own API in case of needed changes, I'm moving this to an
internal package.