feat: support conditional expressions in vectorized map (#5017)
* feat: allow cond expressions to vectorize per semantic graph

* chore: make generate

* feat: add runtime support for vectorized conditional expressions

Updates the Go side to support vectorized conditionals.
Adds new template-generated code to handle the arrow array building code
as well as updates to the `values` package which acts as an entry point
to help delegate to the correct typed API based on the MonoType of the
incoming vectors.

* test: show that conditional expressions are vectorizable

Currently the Go unit tests that actually process inputs are failing due
to a type conflict on the `test` portion of the expression,
ex: `bool != v[bool]`.

Not sure why I didn't bump into this earlier -- manual testing via
scripts all seemed to work fine as is. Watching in the debugger showed
execution hitting all the expected (vectorized) code paths.

The "Logical Ops" PR added a `typ` field to the expression node to allow
us to alter the monotype during vectorization, so I'll do the same and
see if that fixes things.

* fix: test the vectorized fn instead of the original

Fixes an issue where the tests for vectorized conditionals broke with a
type conflict on the `test` part of the expression (`bool` vs
`v[bool]`).

For whatever reason, it seems like we may have been running the
row-based version of the function prior to this diff.

Now, when a test case sets `vectorizable` to `true`, we replace the
function expression with the vectorized version before passing it to
`Compile`.

* test: add acceptance tests for vectorized conditionals

The test cases try to cover the range of types we support, while also
targeting constants in the consequent/alternate positions to verify vec
repeat plays well.

Right now, the tests exercising vec repeat fail! Fixes pending.

* fix: add missing vec repeat handling for vectorized conditionals

Already covered were cases where `test` was const, as well as when
EITHER `consequent` OR `alternate` were const, but not BOTH.

For cases where all 3 are const, the receiver of of the return value of
the conditional expression will get the vec repeat from whichever branch
was selected (this case was already handled).

The panic from the type conflict between Time/Int is still to resolve.

* fix: explicitly cast Time to Int when handling vec repeats

* chore: comment cleanup

* chore: make generate

* refactor: add `vectorizedConditionals` feature flag

Adds a new flux feature flag, propagating it down into the rust analyzer
code so we can use it to activate or disable the new code path.

Tests have been updated to enable the flag, but it is off by default.

* chore: make generate

* refactor: allow nils in vectorized test case inputs

Lightly modified version of work done in this (currently draft) PR:
https://github.com/influxdata/flux/pull/5015

* test: add test cases for null test, consequent, alternate

* chore: tidy up tmpl

* refactor: simplify null checks for test, consequent, alternate