fix: check bounds using rune array instead of string len in substring
Inputs that contain non-breaking spaces (NBSP), and assumedly a host of
other unicode characters caused `strings.substring` to panic when it
tried to index into an array of runes taken from a (seemingly larger)
string.

Failures looked like:
```
panic: runtime error: slice bounds out of range [:33] with capacity 32
```

This only seemed to panic when the inputs were "of a certain size" and
I'm not sure why that would be. Perhaps there are distinct thresholds
where the input causes the capacity to be just close enough to the
actual size of the rune count, but less than the string length in bytes.
Not sure.

The testcase added used an input seen in the wild and panicked prior to
the code change to substring.