refactor: avoid building array if not needed
In the vectorized ToFloatConv a base case is when it receives an array
of floats (no conversion is needed). In this situation we can return the
input as-is.

Since the rest of the inputs will return a brand new arrays, the
expectation seems to be the caller would release the input. To mimic
this behavor for the reuse case, we tick up the refcount by calling
`.Retain()` before handing the array back to the caller.