feat: rewrite calls to float as _vectorizedFloat
For functions matching our point of entry pattern, `(r) => A`
rewrite calls to `float()`, even those in the stdlib.

This rewrite is naive in that it just plops a new identifier in place of
the old. Unsure if there's a way to do a proper lookup for the real
Symbol or not.