fix: restore int return value for `pagerduty.sendEvent`
A recent change updated the HTTP call inside sendEvent to use requests
instead of the old http package which accidentally introduced a breaking
change by switching the return from an int (status code) to a record
(http response).

This diff will rename the new (response object) version of `sendEvent`
to `_sendEvent` so it can be used in `endpoint()`. `sendEvent` will then
call `_sendEvent` and return only the `statusCode` field to restore the
original signature.

Refs: <https://github.com/influxdata/flux/pull/5006>