fix: fixup int/uint comparisons where int is < 0
Both the vectorized and row-based code paths for equality ops are now
"correctly" handling cases where the int side of the up is less than
zero.

The rule we're currently following is like:

- when int is on the left and sub-zero, it CANNOT be larger or equal
- when int is on the right and sub-zero, it CANNOT be smaller or equal

Implied in the above is the opposite side is a uint, which therefore
must be zero or larger.

With the row-based code path fixed up, and vectorized made to match, the
new acceptance tests for vectorized equality ops have been updated to
align with all of the above.