chore: Remove the version range from lsp-types
This does not work as I would hope (see https://github.com/rust-lang/cargo/issues/9029) so we need
to restrict this to just the same version that lspower uses (for the sake of the LSP). This is bad in its own way though as lspower does not adhere to semver which can cause its own breakages.

Personally I'd like use to remove the `From` implementations we provided here and let the LSP have its own, explicit conversions so we could avoid this churn :shrug:.