test: Add debug.null()
Allow null values to be created explicitly in flux code. We could always create nulls by using `csv.from` however that format is pretty hard to read compared to `array.from` but since the latter could not create streams with nulls we were forced to use `csv.from`. With this we can use `array.from` in cases where nulls are needed in the test.