fix(build): update buildinfo doc comments to match latest gofmt
ci throws an error comparing buildinfo.gen.go file when my local go version is 1.19 and ci is on previous
version.
This is because gofmt is changed in 1.19 and now reformats doc comments with extra line when there's no space between // and comments in the format a:b. https://tip.golang.org/doc/go1.19#go-doc So reformatted the comments itself to work with current and previous versions.