feat: add support for vectorized binary equality ops (#5042)
* feat: add vectorized impls for binary equality ops on Go side

* test: add Go-side tests for vectorized binary equality ops

* feat: vectorize binary equality ops on rust side

* feat: add `vectorizedEqualityOps` feature flag

* chore: make generate

* refactor: special case Time, rename ops, flesh out acceptance tests

While writing acceptance tests, I found:
- Time needed to be special cased
- The error messages for unsupported ops/types were not great with the
  short op names I'd used (ie, Lt, Lte, Eq...)

While patching in the support for Time, I also renamed the ops to have
more human readable names, Equal instead of Eq and so on.

Acceptance tests now cover the full range of supported types, as well as
using vec repeats.

* chore: make generate

* chore: fixup flag description

* chore: make generate

* feat: support implicit casts for numeric types

* chore: add comment about redundant const casts

Not particularly interested in making the template harder to read just
yet. Will write tests first then worry about removing the redundant
casts.

* test: add tests for casts

Note that the int/uint gt and gte casts seem to have the wrong answer
here, and in the row-based version of the op. Will need to refactor the
template and row-based impl to get this to pass.

* fix: fixup int/uint comparisons where int is < 0

Both the vectorized and row-based code paths for equality ops are now
"correctly" handling cases where the int side of the up is less than
zero.

The rule we're currently following is like:

- when int is on the left and sub-zero, it CANNOT be larger or equal
- when int is on the right and sub-zero, it CANNOT be smaller or equal

Implied in the above is the opposite side is a uint, which therefore
must be zero or larger.

With the row-based code path fixed up, and vectorized made to match, the
new acceptance tests for vectorized equality ops have been updated to
align with all of the above.

* chore: make generate