refactor: general refactors around removing unused functions and types (#5094)
* refactor: remove spec json marshal and unmarshal functions

The spec marshaling and unmarshaling functions are unused and don't
work. Removing the code related to this.

* refactor: remove all calls to RegisterOpSpec

This removes all calls to `RegisterOpSpec` with the intent to deprecate
and remove the function.

This function doesn't do anything in the current version of Flux. A long
time ago, it was possible to serialize and deserialize the flux spec
which was generated from the interpreter.

We removed this ability from everything except for the interpreter
because it was unreliable and led to bugs. Since this code was only
useful for the serialization and deserialization, and this is long
unsupported and doesn't work, I'm finally removing the setup function
that enabled this capability since it doesn't work and has led mostly to
confusion about what's necessary when creating a source/transformation.

* refactor: remove unused spec format functions

* refactor: remove the unused priority type