fix(stdlib): testing.shouldError uses a regex instead of a string (#5099)
The changes to the test framework necessitated this change. Since the
changes to the test framework now include the full path, we can
potentially run the tests with different paths being passed into the
test framework.

This meant that `shouldError` wasn't able to capture errors because we
can't do a single string that matches the error due to file names.

In general, it's better to do this kind of comparison with regexes so
I've changed it to use a regex.