feat: rewrite calls to `float()` as `_vectorizedFloat()` (#5048)
* feat: rewrite calls to float as _vectorizedFloat

For functions matching our point of entry pattern, `(r) => A`
rewrite calls to `float()`, even those in the stdlib.

This rewrite is naive in that it just plops a new identifier in place of
the old. Unsure if there's a way to do a proper lookup for the real
Symbol or not.

* test: confirm calls to float can vectorize from Go

* test: acceptance test for _vectorizedFloat

Bool literals were not included in the vec repeat test since they are
still not eligible for vectorization. Ref: #4997

* feat: add feature flag for `vectorizedFloat`

* chore: remove fixed fixme

* chore: define builtin float in test instead of loading stdlib

* chore: move float builtin def to analyzer Environment

Feels more official this way.

* chore: make generate