fix: sort columns when printing group key
The flux pivot() function uses a string representation of
the group key to keep track of what columns it has seen.
Sometimes the columns of the group key can appear in a different order
even if the set of columns and their values are the same. when this
happens, pivot() panics as it uses a map that assumes all group
keys have a canonical representation.

The fix here is to make `groupKey.String()` use the sorted order
of the group columns when returning its result.

Fixes #5104.