fix: handle null test, consequent, alternate in vectorized conditional
Initial work for vectorized conditionals neglected to cover cases where
the inputs for `test`, `consequent`, and `alternate` could be invalid
(ie, referencing fields which could not be statically verified and
missing at runtime).

This diff aims to handle these cases by:
- treating a null `test` the same as "all false" by returning
  `alternate` directly.
- changing the constant (vec repeat) code paths to accept pointers to
  the various primitive types so they can be optionally null.
- short circuiting for cases where both `consequent` and `alternate` are
  null (since the output will be null, regardless of the values in `test`).

The choice to adjust the constant handling code path to also handle the
null case was largely motivated by wanting to avoid adding yet another
factor to drive combinations of function signatures.

This would have required a large expansion of functions in the `array`
package, as well as higher complexity to be able to delegate to the new
functions from the `values` side.

We stated with needing (per type):
- vec, const
- const, vec
- vec, vec
- const, const
By going this way, we avoid adding (per type):
- vec, null
- null, vec
- const, null
- null, const

Furthermore, the logic required for nulls in either position is more or
less identical to the const pattern, with the addition of a null check
of the pointer.

Refs:
- <https://github.com/influxdata/flux/issues/4601>
- <https://github.com/influxdata/flux/pull/5017>