fix: Don't treat errors in SQL syntax as internal
Since the query is written by a user, it is not an internal error and the user should get the error reported back to them so they may correct the issue.

Added `shouldErrorWithCode` to allow tests to check the error code and tweaked `catch` to supply the code.

`shouldErrorWithCode` should maybe be in internal or experimental instead? I don't think we want to expose the error code to users.