feat(function): update function library to have its own arguments struct
The function package now uses its own arguments struct. This is to
eventually remove the `interpreter.Arguments` interface. An interface is
a poor choice for that type because interfaces with that many methods
cannot be changed or adapted easily.

At some point, the function package will be used for all situations
instead of the raw register methods.