test: Correct the value of _stop and now in shapeData (#5133)
* test: Correct the value of _stop and now in shapeData

`_stop` were earlier than some of the values of `_time` and `now`, which is used as the default for `range` were different
from the `_stop` specified in the CSV files which causes differences when using `diff` to compare. Correcting and normalizing these
fixes the problem when testing against influxDB.

I am not quite sure why it still works when just running the test locally though. I guess `range` never updates `_stop` in that case so it is always the default?

* chore: Fix ShapedataWithFilter