fix: inconsistent runtime typing for `logicalVectorEvaulator`
When a logical op appears in the `test` position for a conditional,
we see:

```
cannot use test of type vector in conditional expression; expected boolean
```

The runtime typing for `logicalVectorEvaluator` was bool
rather than a vector of bool which would cause the conditional evaluator
to incorrectly select the non-vectorized implementation (it expects
test, consequent and alternate to all be vectors in order to use the
vectorized version).

By updating the `Type()` method for `logicalVectorEvaluator` to be a
vector of bool, the `conditionalVectorEvaluator` is able to be used.