fix: treat null inputs to `vectorAdd`, `vectorOr` as false
Formerly, when the input to vectorized logical ops was a null, we'd see:

```
runtime error: cannot use vectorized or operation on non-vector invalid
```

The precedent set by the row-based logical ops is to treat incoming
nulls as false.

This diff extracts the input validation to a separate function which is
responsible for transitioning the incoming `values.Value` to a `values.Vector`
giving it the chance to intercept the problem null case and convert them
to a new vec repeat holding a boolean false.