feat: add logicalStrictNullEvaluator
A new version of the logical evaluator which matches the spec.
The first thing it does, is check for a feature flag and will delegate
to the original version when not enabled.

This is not ideal. It would be better to branch higher up in the
compiler, but we don't have easy access to the context to read the
flagger at that level.

BREAKING CHANGE: nulls propagate with the `strictNullLogicalOps` flag on