refactor: fixup row-based to add short circuiting back
As I realized these logical ops are commutative I initially thought it
meant short circuiting wouldn't be possible (since we'd need to know the
null positions on both sides) but this was not true.

We can skip evaluating the right side of the op when we see the terminal
cases of a `true` for OR, and `false` for AND. Nulls still need to be
considered carefully case by case when we resort to looking at the right-hand
side, but we can still short circuit.