fix: Prevent the metadata map from being concurrently mutated
By protecting with an `RWMutex`. I haven't been able to reproduce the crash but given the stacktrace it seems obvious enough that it is because `Metadata` is stored in the `Context` for which there is only one per query. Since queries run on multiple threads we must ensure that multiple threads can't mutated it concurrently.

Fixes #5157