docs: update design section and merge it with the old development plan
I've also changed some of my opinions on the strictness of SSA form. I
don't really think it matters anymore but it also may be nice to have.
It seems more like an implementation detail so I removed it.

I've also changed my opinion on where the VM should be implemented. I
think, even if we make a custom VM, Rust is a better place. Rust has
more consistent performance that's not arbitrarily affected by a GC and
makes a better choice as a host language for a VM.

I think we can use C FFI to interact with parts that need to interact
with Go.