fix(stdlib): sort limit skips chunks with no rows
The sort limit code did not skip chunks with no rows which caused it to
attempt to compare invalid chunks with each other. This skips those rows
but still ensures that the schema is created for them.

This is similar to the same problem that was identified in #4701. That
PR only addressed when the last or only chunk was zero while this one
ensures that the aggregation step also handles this situation properly.