fix: Prevent the metadata map from being concurrently mutated (#5209)
* fix: Prevent the metadata map from being concurrently mutated

By protecting with an `RWMutex`. I haven't been able to reproduce the crash but given the stacktrace it seems obvious enough that it is because `Metadata` is stored in the `Context` for which there is only one per query. Since queries run on multiple threads we must ensure that multiple threads can't mutated it concurrently.

Fixes #5157

* refactor: Split Metadata into synchronous and non-synchronous variants

Simplifies things significantly as the parts that wants to use `Metadata` as a value still work and we can use the synchronous only where it is needed.