fix: propagate null inputs to vectorized float()
The vectorized version of the `float()` conversion builtin correctly handled
the case where elements in the incoming vector were null, but didn't
cover the case where the vector itself is null.

By short-circuiting before we do any array processing, we mimic the
standard `float()` behavior of producing an untyped null.