chore: re-flag strict null logical evaluator
A new version of our logical evaluator was introduced in an effort to
re-align the lang with the spec regarding nulls. Ref:
<https://github.com/influxdata/flux/pull/5192>

The "strict null" flavor of the logical evaluator was
bypassed as part of a perf regression investigation. Ref:
<https://github.com/influxdata/flux/pull/5221>

The thought was checking the feature flag with each eval was adding
overhead and raising query latency, but with a susequent refactor
(<https://github.com/influxdata/flux/pull/5222>) we are now able to
check the flag earlier, reducing the overhead.