feat(dynamic): runtime support for member expressions, remove index support (#5258)
* test: add tests for valid/invlaid/undefined/oob member and index access

* feat: implement runtime member access for dynamic

* refactor: remove support for arbitrary index exprs on dynamic

Initially we thought to split activities performed on dynamic values
into 2 categories:

- traversal
- conversion (ex: casting)

Our objective was roughly to have users cast values at the leaves of the
data structure, while reducing friction as they descend into it.

When we decided to add `asArray` this undercut the need for treating
indexing as a traversal activity. Instead folks can convert to an array
and then index into it.

The fallibility of the array cast is still something we need to reduce
friction on, but likely this will happen via new cast functions. We
don't need to enable this arbitrarily at this stage I don't think.

* test: add deep field (null) access on dynamic

* chore: make generate