fix(stdlib): array touint test utilized conversion behavior incorrectly (#5274)
The `array.toUInt()` test utilized conversion behavior and assumed that
conversion behavior was the same for floats to uints as ints to uints.

Floats to uints is implemented differently than ints to uints especially
on the ARM platform. On ARM, a negative float converts to zero. While we
may want to change Flux to define a specific behavior for this
conversion in the long run, this is just going to fix the test so it
consistently uses floats for now so the test passes appropriately on all
architectures.