refactor(dynamic): member expr should error when inner is not obj
Member expressions on dynamic values started out very loose.
Originally, any dynamic would just hand back a `dynamic(null)` when the
property lookup failed. This means getting good feedback from the
runtime was much harder as nulls could just bleed through.

This diff aims to ensure we give better feedback - especially for cases
where the dynamic inner value is some non-object (int, float, null, and
so on).
In these situations a property lookup would never make sense. This
indicates the underlying data structure is probably unexpected to the script
author and updates to the program will be needed.

Overall this should be a safer initial implementation at the language
level since it is "more regular" and less exotic. We can add API to
improve the developer experience without much sacrifice.

For cases where the inner value _is an object_ but the property doesn't
exist, we will continue to hand back a `dynamic(null)` as I believe this
would be expected. The key thing is to error when the overall shape of
the underlying data has drifted away from the original expectations.

N.B. this change means the `dynamic_cast_from_json_deep` test had to be
updated to use `exists` to dance around the nulls. It also seems like
users will need to leverage `debug.null()` for alternate cases which we
might not really like since it's "internal."

This likely means <https://github.com/influxdata/flux/issues/5287> or
something in a similar space will be an important workflow element,
ergonomically moving forward.