chore(dev): first attempt at a vscode devcontainer for the project
This should create a standard environment for contributors to load up
the project and use vscode to modify it and have a usable environment
for compiling and running the program.

It still leaves a lot to be desired. For whatever reason, loading the
project takes forever. `rust-analyzer` takes a very long time and the go
tools aren't presently installed by default. We can iterate on the
container to make it quicker to load and a more desirable environment to
work in.