fix: shouldError should error when no error occurs
When `catch` did not get an error the `msg` field would be null, which would make the regexp match operator to just propagate `null` instead of trying to evaluate. Since `null` gets translated to false in the filter we would get an empty stream, causing us to not emit an error.

Fixes https://github.com/influxdata/flux/issues/5289